+++ src/controllers/recipeController.js
@@ -1,14 +1,18 @@
-const { success, error, validation } = require("../utils/helpers/baseResponse");
+const { success, successPage } = require("../utils/helpers/baseResponse");
 const { Validator } = require('../utils/helpers/validator');
 const { HandlerException } = require('../utils/helpers/errorHandler');
 const { GetConfigPagination } = require('../utils/helpers/paginatorInit');
 
 const statusCode = require('../utils/helpers/statusCode');
 const recipeService = require('../services/recipeService');
+const { MapListRecipes } = require("../models/responses/recipe/getRecipe");
 
 const postRecipe = (req, res) => {
     try {
         var body = req.body;
+
+        Validator.ValidateId(body.userId, "El id del usuario es inválido");
+        Validator.ValidateMatchTokenUserId(body.userId, req.data);
 
         return recipeService.postRecipe(body).then(recipe => {
             res
@@ -27,7 +31,28 @@ const getAllRecipe = (req, res) => {
         return recipeService.getAllRecipe(pagination).then(recipes => {
             res
                 .status(statusCode.OK)
-                .json(success("OK", recipes.data, statusCode.OK, pagination.header, recipes.totalPage));
+                .json(successPage("OK", recipes.data, statusCode.OK, pagination.header, recipes.totalPage));
+        });
+    } catch (e) {
+        HandlerException(e, res);
+    }
+};
+
+const getAllRecipeForAdmin = (req, res) => {
+    try {
+        var pagination = GetConfigPagination(req);
+        var status = req.params.status;
+
+        return recipeService.getAllRecipe(pagination, status).then(recipes => {
+            if (recipes == null) {
+                res
+                    .status(statusCode.OK)
+                    .json(success("Sin resultados", [], statusCode.OK));
+            } else {
+                res
+                    .status(statusCode.OK)
+                    .json(successPage("OK", MapListRecipes(recipes.data), statusCode.OK, pagination.header, recipes.totalPage));
+            }
         });
     } catch (e) {
         HandlerException(e, res);
@@ -41,80 +66,94 @@ const getRecipeById = (req, res) => {
         Validator.ValidateId(recipeId, "El id de la receta es inválido");
 
         return recipeService.getRecipeById(recipeId).then(recipe => {
-            res
-                .status(statusCode.OK)
-                .json(success("OK", recipe, statusCode.OK));
+            if (recipe == null) {
+                res
+                    .status(statusCode.OK)
+                    .json(success("No se encontró un usuario con el id proporcionado", null, statusCode.OK));
+            } else {
+                res
+                    .status(statusCode.OK)
+                    .json(success("OK", recipe, statusCode.OK));
+            }
         });
     } catch (e) {
         HandlerException(e, res);
     }
 };
 
-const updateRecipe = (req, res) => {
+const updateRecipe = async (req, res) => {
     try {
         var body = req.body;
         var recipeId = req.params.recipeId;
 
         Validator.ValidateId(recipeId, "El id de la receta es inválido");
+        Validator.ValidateId(body.userId, "El id del usuario es inválido");
+        Validator.ValidateMatchTokenUserId(body.userId, req.data);
+
+        var recipe = await recipeService.getRecipeById(recipeId);
+
+        if (!recipe) {
+            return res
+                .status(statusCode.OK)
+                .json(success("No se encontró una receta con el id proporcionado o está desactivada", null, statusCode.OK));
+        }
+        Validator.ValidateOwner(recipe.user.id, body.userId);
 
         return recipeService.updateRecipe(recipeId, body).then(updatedRow => {
-            if (updatedRow == null) {
-                res
-                    .status(statusCode.OK)
-                    .json(success("Receta a actualizar no encontrada", null, statusCode.OK));
-            } else {
-                res
-                    .status(statusCode.OK)
-                    .json(success("OK", updatedRow, statusCode.OK));
-            }
+            res
+                .status(statusCode.OK)
+                .json(success("OK", updatedRow, statusCode.OK));
         });
     } catch (e) {
         HandlerException(e, res)
     }
 };
 
-const deleteRecipe = (req, res) => {
+const deleteRecipe = async (req, res) => {
     try {
         var recipeId = req.params.recipeId;
 
         Validator.ValidateId(recipeId, "El id de la receta es inválido");
+
+        var recipe = await recipeService.getRecipeById(recipeId);
+
+        if (!recipe) {
+            return res
+                .status(statusCode.OK)
+                .json(success("No se encontró una receta con el id proporcionado o ya se encuentra desactivada", null, statusCode.OK));
+        }
 
         return recipeService.deleteRecipe(recipeId).then(deletedRow => {
-            if (deletedRow == null) {
-                res
-                    .status(statusCode.OK)
-                    .json(success("Receta a desactivar no encontrada", null, statusCode.OK));
-            } else {
-                res
-                    .status(statusCode.OK)
-                    .json(success("OK", deletedRow, statusCode.OK));
-            }
+            res
+                .status(statusCode.OK)
+                .json(success("OK", deletedRow, statusCode.OK));
         });
     } catch (e) {
         HandlerException(e, res)
     }
 };
 
-const reactivateRecipe = (req, res) => {
+const reactivateRecipe = async (req, res) => {
     try {
         var recipeId = req.params.recipeId;
 
         Validator.ValidateId(recipeId, "El id de la receta es inválido");
 
+        var recipe = await recipeService.getRecipeById(recipeId, 0);
+        if (!recipe) {
+            return res
+                .status(statusCode.OK)
+                .json(success("No se encontró una receta con el id proporcionado o ya se encuentra activa", null, statusCode.OK));
+        }
+
         return recipeService.reactivateRecipe(recipeId).then(reactivatedRow => {
-            if (reactivatedRow == null) {
-                res
-                    .status(statusCode.OK)
-                    .json(success("Receta a reactivar no encontrada", null, statusCode.OK));
-            } else {
-                res
-                    .status(statusCode.OK)
-                    .json(success("OK", reactivatedRow, statusCode.OK));
-            }
+            res
+                .status(statusCode.OK)
+                .json(success("OK", reactivatedRow, statusCode.OK));
         });
     } catch (e) {
         HandlerException(e);
     }
 };
 
-module.exports = { postRecipe, getAllRecipe, getRecipeById, updateRecipe, deleteRecipe, reactivateRecipe };
+module.exports = { postRecipe, getAllRecipe, getAllRecipeForAdmin, getRecipeById, updateRecipe, deleteRecipe, reactivateRecipe };
